# swagger-docker/dist/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "ACME project is a MVP project",
    "version" : "v1",
    "title" : "ACME Rest Documentation",
    "termsOfService" : "http://invillia.com/terms",
    "contact" : {
      "name" : "Support",
      "url" : "http://invillia.com.br",
      "email" : "support@invillia.com"
    },
    "license" : {
      "name" : "Invillia License 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/api/store" : {
      "get" : {
        "tags" : [ "Store Services" ],
        "summary" : "",
        "description" : "Returns all stores",
        "operationId" : "getStores",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved stores",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Store Services" ],
        "summary" : "",
        "description" : "Save store",
        "operationId" : "saveStore",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/StoreInputDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/StoreOutputDTO"
            }
          },
          "201" : {
            "description" : "Store successfully saved",
            "schema" : {
              "$ref" : "#/definitions/StoreOutputDTO"
            }
          },
          "403" : {
            "description" : "Store already saved",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/store/getById/{store-id}}" : {
      "get" : {
        "tags" : [ "Store Services" ],
        "summary" : "",
        "description" : "Gets the store by id",
        "operationId" : "getStoreById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "store-id",
          "in" : "path",
          "description" : "Store Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the store",
            "schema" : {
              "$ref" : "#/definitions/StoreOutputDTO"
            }
          },
          "404" : {
            "description" : "The store id was not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/store/getByName/{store-name}}" : {
      "get" : {
        "tags" : [ "Store Services" ],
        "summary" : "",
        "description" : "Gets the store by name",
        "operationId" : "getStoreByName",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "store-name",
          "in" : "path",
          "description" : "Store name",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the store",
            "schema" : {
              "$ref" : "#/definitions/StoreOutputDTO"
            }
          },
          "404" : {
            "description" : "The store name was not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/store/{store-id}" : {
      "post" : {
        "tags" : [ "Store Services" ],
        "summary" : "",
        "description" : "Updates the store by id",
        "operationId" : "updateStore",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "store-id",
          "in" : "path",
          "description" : "Store id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "StoreDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/StoreInputDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The store was updated sucessfully",
            "schema" : {
              "$ref" : "#/definitions/StoreOutputDTO"
            }
          },
          "404" : {
            "description" : "The store was not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "ErrorDTO" : {
      "type" : "object",
      "properties" : {
        "error" : {
          "type" : "string"
        }
      }
    },
...
    "StoreInputDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        }
      }
    },
    "StoreOutputDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        }
      }
    }
  }
}